order vuex store moved to module
