+++ core/modules/order/store/actions.ts
@@ -1,9 +1,9 @@
 import Vue from 'vue'
-import * as types from '../../mutation-types'
-import { ValidationError } from '../../lib/exceptions'
+import * as types from './mutation-types'
+import { ValidationError } from '@vue-storefront/store/lib/exceptions'
 import { currentStoreView } from '@vue-storefront/store/lib/multistore'
 import { ActionTree } from 'vuex'
-import RootState from '../../types/RootState'
+import RootState from '@vue-storefront/store/types/RootState'
 import OrderState from './types/OrderState'
 const Ajv = require('ajv') // json validator
