+++ core/modules/order/store/mutation-types.ts
@@ -1,0 +1,3 @@
+export const SN_ORDER = 'order'
+export const ORDER_PLACE_ORDER = SN_ORDER + '/PLACE_ORDER'
+export const ORDER_PROCESS_QUEUE = SN_ORDER + '/PROCESS_QUEUE'
